Delete READme.md
This repository contains Projects performed during Google All India Scholarship Program. The Google Scholarship Program is a Google Initiative of up-skilling Indians as part of Digital India Program. The Program is conducted by Udacity in partnership with Google. I was one of the Indians selected of the total 10,000 skills after my application was evaluated by Udacity and Google. The Scholarship Program has 4-6 projects adhering to strict to timelines.